Issue #72: Simplify export of V8Js properties to JavaScript.
Instead of using special JavaScript accessors for the `PHP` object,
we just reuse the standard PHP object wrapping.  To control which properties
are visible to JavaScript, we use a separate PHP V8UserProperties object
for the JavaScript-visible stuff.

We need to use __get/__set/etc magic methods (instead of the object
handler equivalents) in order to make things work properly with derived
classes.