Test results changed in revised issue #72 patch, w/ read-only export.
There is a minor change to how properties of V8Js subclasses are exposed to
JavaScript: subclass properties are *not* visible to JS by default.

We preserve the read-only status of PHP objects exposed via V8Js properties
by adding a hidden read-only flag to the V8 wrappers.  PHP objects exposed
to JS by other means (for example, return values of method invocations)
are not read-only.